i added stuff from 3.6 and 3.7 and more
the pig hammer, wizard wand, skeleton sword, and redstone cannon have been added. along with redstone bullets for the redstone cannon and a fix to the broken uzi. winter island was also added. some music stuff and misc was fixed/changed. re-implemented the facebook button and made it swap lobbies. and more such as the use of the //troubleshooting: keyword to help if I'm having issues in the future.